+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed:
 - `motionbuilder.setupCodeCompletion` now downloads the latest stubs from [pyfbsdk-stub-generator/generated-stub-files](https://github.com/nils-soderman/pyfbsdk-stub-generator/tree/main/generated-stub-files)
 - The documentation & examples now browses the 2025 docs
+- Extension appdata is now stored in VS Code's user-data folder
 
 ### Fixed:
 - Bug preventing VS Code from attaching to MoBu 2025